move parser tests into subdirectory
