+++ tests/parsers/test_parser_delete_extracts.py
@@ -1,7 +1,7 @@
 import unittest
 
 from tabcmd.parsers.delete_extracts_parser import DeleteExtractsParser
-from .common_setup import *
+from common_setup import *
 
 commandname = "deleteextracts"
