+++ tests/parsers/test_parser_delete_site.py
@@ -1,24 +1,22 @@
 import unittest
+from tabcmd.parsers.delete_site_parser import DeleteSiteParser
+from common_setup import *
 
-import argparse
-from tabcmd.parsers.reencrypt_parser import ReencryptExtractsParser
-from .common_setup import *
-
-commandname = "reencryptextracts"
+commandname = "deletesite"
 
 
-class ReencryptExtractsParserTest(unittest.TestCase):
+class DeleteSiteParserTest(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
         cls.parser_under_test, manager, mock_command = initialize_test_pieces(commandname)
-        ReencryptExtractsParser.reencrypt_extracts_parser(manager, mock_command)
+        DeleteSiteParser.delete_site_parser(manager, mock_command)
 
-    def test_reencrypt_extract_parser_optional_arguments(self):
+    def test_delete_site(self):
         mock_args = [commandname, "site-name"]
         args = self.parser_under_test.parse_args(mock_args)
         assert args.site_name == "site-name", args
 
-    def test_reencrypt_extract_parser_missing_all_args(self):
+    def test_delete_site_required_name_none(self):
         mock_args = [commandname]
         with self.assertRaises(SystemExit):
             args = self.parser_under_test.parse_args(mock_args)
