+++ tests/parsers/test_parser_get_url.py
@@ -2,7 +2,7 @@ import unittest
 
 import argparse
 from tabcmd.parsers.get_url_parser import GetUrlParser
-from .common_setup import *
+from common_setup import *
 
 commandname = "listsites"
