+++ tests/parsers/test_parser_login.py
@@ -2,7 +2,7 @@ import unittest
 
 import argparse
 from tabcmd.parsers.login_parser import LoginParser
-from .common_setup import *
+from common_setup import *
 
 commandname = "login"
